nonview: Copied Cache and WWW (new) from sri_lanka

+++ src/nonview/base/Cache.js
@@ -1,46 +1,16 @@
-const N_MAX_RETRY = 5;
 export default class Cache {
-  constructor(cacheKey) {
-    this.cacheKey = cacheKey;
-  }
-
-  clear() {
-    localStorage.setItem(this.cacheKey, "");
-  }
-
-  set(item) {
-    for (let i = 0; i < N_MAX_RETRY; i++) {
-      try {
-        localStorage.setItem(this.cacheKey, JSON.stringify(item));
-        return true;
-      } catch {
-        localStorage.clear();
-      }
+  static async get(cacheKey, asyncFallback) {
+    const hotItem = localStorage.getItem(cacheKey);
+    if (hotItem) {
+      return JSON.parse(hotItem);
     }
-    return false;
-  }
 
-  getHot() {
-    const hotJSON = localStorage.getItem(this.cacheKey);
-    if (Cache.isValidHotItem(hotJSON)) {
-      return JSON.parse(hotJSON);
-    }
-    throw Error("Item not in cache: " + this.cacheKey);
-  }
-
-  async get(asyncFallback) {
+    const coldItem = await asyncFallback();
     try {
-      return this.getHot();
-    } catch (e) {}
-
-    const cold = await asyncFallback();
-    this.set(cold);
-    return cold;
-  }
-
-  static isValidHotItem(itemJSON) {
-    return (
-      itemJSON && itemJSON !== "" && itemJSON !== null && itemJSON !== "null"
-    );
+      localStorage.setItem(cacheKey, JSON.stringify(coldItem));
+    } catch (QuotaExceededError) {
+      localStorage.clear();
+    }
+    return coldItem;
   }
 }
